+++ CHANGELOG.md
@@ -10,3 +10,9 @@
 ## [0.0.3] - 19.04.2020
 
 * Fixing the problem with padding
+
+## [0.1.0] - 06.05.2020
+
+* Changing background color of the banner
+* Chnaging text style
+* Canging border radius